requirements.txt
PyYAML
mkdocs
mkdocs-material
markdown-include
anthropic
